refactor: add LockByPartitionFacet and update configurations
