+++ packages/ats/contracts/Configuration.ts
@@ -114,6 +114,7 @@ export const CONTRACT_NAMES = [
   "CorporateActionsFacet",
   "TransferAndLockFacet",
   "LockFacet",
+  "LockByPartitionFacet",
   "AdjustBalancesFacet",
   "BalanceTrackerFacet",
   "BalanceTrackerAdjustedFacet",
