+++ packages/ats/contracts/contracts/constants/resolverKeys.sol
@@ -210,6 +210,9 @@ bytes32 constant _CORPORATE_ACTIONS_SUSTAINABILITY_PERFORMANCE_TARGET_RATE_RESOL
 // keccak256("security.token.standard.lock.resolverKey")
 bytes32 constant _LOCK_RESOLVER_KEY = 0xf1364345b3db5ebe5808f2d2d2aaecb9cdb4fddacad1534033060ebc886fc1e9;
 
+// keccak256("security.token.standard.lock.by.partition.resolverKey")
+bytes32 constant _LOCK_BY_PARTITION_RESOLVER_KEY = 0x9c79f8b10d67860ea033eb8936f04428327ffd8f704f9bb74446bff26c3ddb2a;
+
 // keccak256("security.token.standard.protected.partitions.resolverKey")
 bytes32 constant _PROTECTED_PARTITIONS_RESOLVER_KEY = 0x6d65d2938c05a4d952aff0845c1baa5bea04d4544db74f8b3b26004d1d58d58f;
