+++ packages/ats/contracts/contracts/facets/layer_1/lock/LockFacet.sol
@@ -15,9 +15,7 @@ contract LockFacet is Lock, IStaticFunctionSelectors {
 
     function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
         uint256 selectorIndex;
-        staticFunctionSelectors_ = new bytes4[](14);
-        staticFunctionSelectors_[selectorIndex++] = this.lockByPartition.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.releaseByPartition.selector;
+        staticFunctionSelectors_ = new bytes4[](8);
         staticFunctionSelectors_[selectorIndex++] = this.forceReleaseByPartition.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getLockByPartition.selector;
         staticFunctionSelectors_[selectorIndex++] = this.lock.selector;
@@ -26,10 +24,6 @@ contract LockFacet is Lock, IStaticFunctionSelectors {
         staticFunctionSelectors_[selectorIndex++] = this.getLockCountFor.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getLocksIdFor.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getLockFor.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getLockedAmountForByPartition.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getLockCountForByPartition.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getLocksIdForByPartition.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getLockForByPartition.selector;
     }
 
     function getStaticInterfaceIds() external pure override returns (bytes4[] memory staticInterfaceIds_) {
