+++ packages/ats/contracts/scripts/domain/bondFixedRate/createConfiguration.ts
@@ -117,6 +117,7 @@ const BOND_FIXED_RATE_FACETS = [
   // Advanced Features
   "AdjustBalancesFacet",
   "LockFacet",
+  "LockByPartitionFacet",
   "NominalValueFacet",
   "ProceedRecipientsFacet",
   "ProtectedPartitionsFacet",
