+++ packages/ats/contracts/scripts/domain/bondKpiLinkedRate/createConfiguration.ts
@@ -112,6 +112,7 @@ const BOND_KPI_LINKED_RATE_FACETS = [
   // Advanced Features
   "AdjustBalancesFacet",
   "LockFacet",
+  "LockByPartitionFacet",
   "NominalValueFacet",
   "ProceedRecipientsKpiLinkedRateFacet", // rate-specific: triggers scheduled tasks
   "ProtectedPartitionsFacet",
