+++ packages/ats/contracts/scripts/domain/equity/createConfiguration.ts
@@ -115,6 +115,7 @@ const EQUITY_FACETS = [
   "DividendFacet",
   "DividendSecurityHoldersFacet",
   "LockFacet",
+  "LockByPartitionFacet",
   "NominalValueFacet",
   "ProtectedPartitionsFacet",
   "ScheduledCrossOrderedTasksFacet",
